+++ docs/CHANGE_LOG.md
@@ -28,6 +28,8 @@
 - Fixed the "Turn On/Off Full-Screen Mode" button divider lines in the views not disappearing in mobile mode.
 - Fixed a fault that caused the wrong week to be shown when the setting "startOfWeekDay" is not Monday and the current day is Saturday or Sunday.
 - Fixed the time arrow not allowing the right context menu to be shown when clicked over an event, or empty space.
+- Fixed event resizing not working correctly when the mouse leaves the event being resized.
+- Fixed the event resize tracking timer always firing in the background.
 
 <br>